+++ apps/backoffice-v2/src/common/components/molecules/DocumentTracker/hooks/useDocumentTracker.tsx
@@ -1,115 +1,21 @@
-import {
-  ctw,
-  DropdownMenuItem,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
-  DropdownMenu,
-} from '@ballerine/ui';
+import { Button, ctw } from '@ballerine/ui';
 import { useQueryClient } from '@tanstack/react-query';
 import { useCallback, useState } from 'react';
 import { titleCase } from 'string-ts';
 
-import { Button } from '@/common/components/atoms/Button/Button';
-import { Input } from '@/common/components/atoms/Input/Input';
-import { Label } from '@/common/components/atoms/Label/Label';
-import { Dialog } from '@/common/components/organisms/Dialog/Dialog';
-import { DialogContent } from '@/common/components/organisms/Dialog/Dialog.Content';
-import { DialogDescription } from '@/common/components/organisms/Dialog/Dialog.Description';
-import { DialogFooter } from '@/common/components/organisms/Dialog/Dialog.Footer';
-import { DialogHeader } from '@/common/components/organisms/Dialog/Dialog.Header';
-import { DialogTitle } from '@/common/components/organisms/Dialog/Dialog.Title';
-import { DialogTrigger } from '@/common/components/organisms/Dialog/Dialog.Trigger';
 import { useRequestDocumentsMutation } from '@/domains/documents/hooks/mutations/useRequestDocumentsMutation/useRequestDocumentsMutation';
 import { useDocumentsTrackerItemsQuery } from '@/domains/documents/hooks/queries/useDocumentsTrackerItemsQuery';
 import { documentsQueryKeys } from '@/domains/documents/hooks/query-keys';
 import { DocumentTrackerItemSchema, TDocumentsTrackerItem } from '@/domains/documents/schemas';
 import { useCurrentCaseQuery } from '@/pages/Entity/hooks/useCurrentCaseQuery/useCurrentCaseQuery';
 import { CommonWorkflowStates } from '@ballerine/common';
-import { DialogClose } from '@radix-ui/react-dialog';
 import z from 'zod';
 import { documentStatusToIcon, Icon } from '../constants';
-import { FilePlus2, MoreVertical, Upload } from 'lucide-react';
-
-type DocumentTrackerItemOptionsProps = {
-  onMarkChange: (reason?: string) => void;
-  isDisabled: boolean;
-};
-
-const DocumentTrackerItemOptions = ({
-  onMarkChange,
-  isDisabled,
-}: DocumentTrackerItemOptionsProps) => {
-  const [reasonValue, setReasonValue] = useState('');
-
-  return (
-    <Dialog>
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <Button
-            variant="outline"
-            size="icon"
-            className="invisible ms-auto text-muted-foreground d-5 focus-visible:visible group-hover:visible aria-disabled:pointer-events-none aria-disabled:cursor-not-allowed aria-disabled:bg-background aria-disabled:opacity-50 data-[state=open]:visible"
-            aria-disabled={isDisabled}
-          >
-            <MoreVertical size={16} />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent align="end" className="px-0">
-          <DropdownMenuItem className="w-full px-8 py-1" asChild>
-            <DialogTrigger asChild>
-              <Button type="button" variant={'ghost'} className="justify-start px-2">
-                <FilePlus2 size={16} className="me-2" />
-                Request from client
-              </Button>
-            </DialogTrigger>
-          </DropdownMenuItem>
-          <DropdownMenuItem className={`w-full px-8 py-1`} asChild>
-            <Button type="button" variant={'ghost'} className="justify-start px-2">
-              <Upload size={16} className="me-2" />
-              Upload
-            </Button>
-          </DropdownMenuItem>
-        </DropdownMenuContent>
-      </DropdownMenu>
-      <DialogContent className="px-16 py-12 sm:max-w-2xl">
-        <DialogHeader>
-          <DialogTitle className="mb-4 text-2xl">Request document from the client</DialogTitle>
-          <DialogDescription className="text-base text-primary">
-            By clicking the &quot;Mark for Request&quot;, the document will be marked as requested.
-            <br />
-            Once marked, you can use the &quot;Request&quot; button button at the top of the
-            documents list to send an email to the customer, asking to upload all of the documents
-            you have marked as needed.
-          </DialogDescription>
-        </DialogHeader>
-
-        <Label htmlFor="reason" className="my-2 font-bold">
-          Reason (Optional)
-        </Label>
-        <Input
-          value={reasonValue}
-          onChange={e => setReasonValue(e.target.value)}
-          id="reason"
-          placeholder="Add reason"
-        />
-        <p>
-          Use the reason input to tell the client why they are required to upload this document. The
-          reason will be visible to the client on the data collection flow on the document uploader
-        </p>
-
-        <DialogFooter>
-          <DialogClose asChild>
-            <Button onClick={() => onMarkChange(reasonValue)}>Mark for Request</Button>
-          </DialogClose>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-};
+import { DocumentTrackerItemOptions } from '../components/DocumentTrackerItemOptions/DocumentTrackerItemOptions';
+import { X } from 'lucide-react';
 
 export const useDocumentTracker = ({ workflowId }: { workflowId: string }) => {
-  const { data: documentTrackerItems, isLoading: isLoadingDocuments } =
-    useDocumentsTrackerItemsQuery({ workflowId });
+  const { data: documentTrackerItems } = useDocumentsTrackerItemsQuery({ workflowId });
 
   const [open, onOpenChange] = useState(false);
   const [selectedIdsToRequest, setSelectedIdsToRequest] = useState<
@@ -126,23 +32,26 @@ export const useDocumentTracker = ({ workflowId }: { workflowId: string }) => {
   });
   const { data: workflow } = useCurrentCaseQuery();
 
-  const onRequestDocuments = () =>
-    requestDocuments({
-      workflowId,
-      documents: selectedIdsToRequest.map(identifier => ({
-        type: identifier.document.type,
-        category: identifier.document.category,
-        issuingCountry: identifier.document.issuingCountry,
-        issuingVersion: identifier.document.issuingVersion,
-        decisionReason: identifier.document.decisionReason,
-        version: identifier.document.version,
-        templateId: identifier.document.type,
-        entity: {
-          id: identifier.entity.id,
-          type: identifier.entity.entityType,
-        },
-      })),
-    });
+  const onRequestDocuments = useCallback(
+    () =>
+      requestDocuments({
+        workflowId,
+        documents: selectedIdsToRequest.map(identifier => ({
+          type: identifier.document.type,
+          category: identifier.document.category,
+          issuingCountry: identifier.document.issuingCountry,
+          issuingVersion: identifier.document.issuingVersion,
+          decisionReason: identifier.document.decisionReason,
+          version: identifier.document.version,
+          templateId: identifier.document.type,
+          entity: {
+            id: identifier.entity.id,
+            type: identifier.entity.entityType,
+          },
+        })),
+      }),
+    [requestDocuments, selectedIdsToRequest, workflowId],
+  );
 
   const getSubItems = useCallback(
     (documentTrackerItem: TDocumentsTrackerItem['business'][number]) => {
@@ -166,11 +75,10 @@ export const useDocumentTracker = ({ workflowId }: { workflowId: string }) => {
       );
       const isSelected = selectedIndex > -1;
 
+      const onUnmark = () => {
+        setSelectedIdsToRequest(prev => prev.toSpliced(selectedIndex, 1));
+      };
       const onMarkChange = (reason?: string) => {
-        if (isSelected) {
-          return setSelectedIdsToRequest(prev => prev.toSpliced(selectedIndex, 1));
-        }
-
         if (status !== 'unprovided') {
           return;
         }
@@ -182,20 +90,29 @@ export const useDocumentTracker = ({ workflowId }: { workflowId: string }) => {
 
       return {
         leftIcon: selectedIndex === -1 ? documentStatusToIcon[status] : Icon.MARKED,
-        rightIcon: (
+        rightIcon: !isSelected ? (
           <DocumentTrackerItemOptions
-            isDisabled={isSelected || status !== 'unprovided'}
+            isDisabled={status !== 'unprovided'}
             onMarkChange={onMarkChange}
           />
+        ) : (
+          <Button
+            variant="outline"
+            size="icon"
+            className="invisible ms-auto text-muted-foreground d-5 focus-visible:visible group-hover:visible aria-disabled:pointer-events-none aria-disabled:cursor-not-allowed aria-disabled:bg-background aria-disabled:opacity-50 data-[state=open]:visible"
+            onClick={onUnmark}
+          >
+            <X />
+          </Button>
         ),
         text: (
           <div className="flex flex-col space-y-0.5">
-            <div className="text-sm font-medium text-gray-900">
+            <span className="text-sm font-medium text-gray-900">
               {titleCase(documentTrackerItem.identifiers.document.category ?? 'N/A')}
-            </div>
-            <div className="text-xs text-gray-500">
+            </span>
+            <span className="text-xs text-gray-500">
               {titleCase(documentTrackerItem.identifiers.document.type ?? 'N/A')}
-            </div>
+            </span>
           </div>
         ),
         itemClassName: ctw('p-1', {
@@ -208,7 +125,6 @@ export const useDocumentTracker = ({ workflowId }: { workflowId: string }) => {
 
   return {
     documentTrackerItems,
-    isLoadingDocuments,
     getSubItems,
     selectedIdsToRequest,
     onRequestDocuments,
