+++ apps/backoffice-v2/src/lib/blocks/hooks/useDocumentBlocks/useDocumentBlocks.tsx
@@ -1,17 +1,14 @@
-import { apiClient } from '@/common/api-client/api-client';
 import { MotionButton } from '@/common/components/molecules/MotionButton/MotionButton';
-import { Method } from '@/common/enums';
 import { checkIsIndividual } from '@/common/utils/check-is-individual/check-is-individual';
 import { ctw } from '@/common/utils/ctw/ctw';
-import { handleZodError } from '@/common/utils/handle-zod-error/handle-zod-error';
 import { useApproveDocumentByIdMutation } from '@/domains/documents/hooks/mutations/useApproveDocumentByIdMutation/useApproveDocumentByIdMutation';
 import { useRejectDocumentByIdMutation } from '@/domains/documents/hooks/mutations/useRejectDocumentByIdMutation/useRejectDocumentByIdMutation';
 import { useRemoveDocumentDecisionByIdMutation } from '@/domains/documents/hooks/mutations/useRemoveDocumentDecisionByIdMutation/useRemoveDocumentDecisionByIdMutation';
+import { useDocumentsAdapter } from '@/domains/documents/hooks/useDocumentsAdapter/useDocumentsAdapter';
 import { useApproveTaskByIdMutation } from '@/domains/entities/hooks/mutations/useApproveTaskByIdMutation/useApproveTaskByIdMutation';
 import { useDocumentOcr } from '@/domains/entities/hooks/mutations/useDocumentOcr/useDocumentOcr';
 import { useRejectTaskByIdMutation } from '@/domains/entities/hooks/mutations/useRejectTaskByIdMutation/useRejectTaskByIdMutation';
 import { useRemoveTaskDecisionByIdMutation } from '@/domains/entities/hooks/mutations/useRemoveTaskDecisionByIdMutation/useRemoveTaskDecisionByIdMutation';
-import { useStorageFilesQuery } from '@/domains/storage/hooks/queries/useStorageFilesQuery/useStorageFilesQuery';
 import { TWorkflowById } from '@/domains/workflows/fetchers';
 import { createBlocksTyped } from '@/lib/blocks/create-blocks-typed/create-blocks-typed';
 import { motionButtonProps } from '@/lib/blocks/hooks/useAssosciatedCompaniesBlock/useAssociatedCompaniesBlock';
@@ -19,174 +16,18 @@ import { useCommentInputLogic } from '@/lib/blocks/hooks/useDocumentBlocks/hooks
 import { checkCanApprove } from '@/lib/blocks/hooks/useDocumentBlocks/utils/check-can-approve/check-can-approve';
 import { checkCanReject } from '@/lib/blocks/hooks/useDocumentBlocks/utils/check-can-reject/check-can-reject';
 import { checkCanRevision } from '@/lib/blocks/hooks/useDocumentBlocks/utils/check-can-revision/check-can-revision';
-import { useDocumentPageImages } from '@/lib/blocks/hooks/useDocumentPageImages';
 import { motionBadgeProps } from '@/lib/blocks/motion-badge-props';
 import { useCaseState } from '@/pages/Entity/components/Case/hooks/useCaseState/useCaseState';
-import { useCurrentCaseQuery } from '@/pages/Entity/hooks/useCurrentCaseQuery/useCurrentCaseQuery';
 import {
   composePickableCategoryType,
-  extractCountryCodeFromDocuments,
   isExistingSchemaForDocument,
 } from '@/pages/Entity/hooks/useEntityLogic/utils';
-import { getDocumentsSchemas } from '@/pages/Entity/utils/get-documents-schemas/get-documents-schemas';
-import { CommonWorkflowStates, StateTag, TDocument, valueOrNA } from '@ballerine/common';
+import { CommonWorkflowStates, StateTag, valueOrNA } from '@ballerine/common';
 import { Button, TextArea } from '@ballerine/ui';
-import { createQueryKeys } from '@lukemorales/query-key-factory';
-import { useQuery } from '@tanstack/react-query';
 import { X } from 'lucide-react';
 import * as React from 'react';
-import { FunctionComponent, useCallback, useMemo } from 'react';
-import { titleCase, toTitleCase } from 'string-ts';
-import { z } from 'zod';
-
-const getDocuments = async ({ entityId, workflowId }: { entityId: string; workflowId: string }) => {
-  const [documents, error] = await apiClient({
-    method: Method.GET,
-    endpoint: `../external/documents/${entityId}/${workflowId}`,
-    schema: z.any(),
-    timeout: 40_000,
-  });
-
-  return handleZodError(error, documents);
-};
-
-const documentsQueryKeys = createQueryKeys('documents', {
-  list: ({ entityId, workflowId }: { entityId: string; workflowId: string }) => ({
-    queryKey: [{ entityId, workflowId }],
-    queryFn: () => getDocuments({ entityId, workflowId }),
-  }),
-});
-
-const useDocumentsQuery = ({ workflowId, entityId }: { workflowId: string; entityId: string }) => {
-  return useQuery({
-    ...documentsQueryKeys.list({ workflowId, entityId }),
-    enabled: !!workflowId && !!entityId,
-  });
-};
-
-export const useDocumentsAdapter = ({
-  entityId,
-  documents: passedDocuments,
-}: {
-  entityId: string;
-  documents: TDocument[];
-}) => {
-  const { data: workflow } = useCurrentCaseQuery();
-  const { data: documentsV2, isLoading: isLoadingDocumentsV2 } = useDocumentsQuery({
-    workflowId: workflow?.id ?? '',
-    entityId,
-  });
-  const { isDocumentsV2 } = workflow?.workflowDefinition?.config ?? {};
-  const generateDocumentTitle = useCallback(
-    ({ category, type, variant }: { category: string; type: string; variant: string }) => {
-      return [valueOrNA(titleCase(category ?? '')), valueOrNA(titleCase(type ?? '')), variant].join(
-        ' - ',
-      );
-    },
-    [],
-  );
-  const identificationDocuments = useMemo(
-    () => passedDocuments?.filter(({ type }) => type === 'identification_document'),
-    [passedDocuments],
-  );
-  const getDocumentPagesBallerineFileIds = useCallback((documents: TDocument[]) => {
-    return (
-      documents?.flatMap(({ pages }) => pages?.map(({ ballerineFileId }) => ballerineFileId)) ?? []
-    );
-  }, []);
-  const documentPages = useMemo(() => {
-    if (isDocumentsV2) {
-      return getDocumentPagesBallerineFileIds(identificationDocuments);
-    }
-
-    return getDocumentPagesBallerineFileIds(passedDocuments);
-  }, [passedDocuments, isDocumentsV2]);
-  const storageFilesQueryResult = useStorageFilesQuery(documentPages);
-  const documentPagesResults = useDocumentPageImages(passedDocuments, storageFilesQueryResult);
-  const documentPagesToDetailsAdapter = useCallback(
-    ({ document, documentIndex }: { document: TDocument; documentIndex: number }) => {
-      return (
-        document?.pages?.map(({ type, fileName, metadata, ballerineFileId }, pageIndex) => {
-          const title = generateDocumentTitle({
-            category: document?.category ?? '',
-            type: document?.type ?? '',
-            variant: metadata?.side,
-          });
-
-          return {
-            id: ballerineFileId,
-            title,
-            fileType: type,
-            fileName,
-            imageUrl: documentPagesResults?.[documentIndex]?.[pageIndex],
-          };
-        }) ?? []
-      );
-    },
-    [documentPagesResults, generateDocumentTitle],
-  );
-  const getDocuments = () => {
-    if (isDocumentsV2) {
-      const adaptedDocumentsV2 =
-        documentsV2?.map(({ decision, decisionReason, issuingCountry, ...document }) => ({
-          ...document,
-          decision: {
-            status: decision === 'revisions' ? 'revision' : decision,
-            reason: decisionReason,
-          },
-          issuer: {
-            country: issuingCountry,
-          },
-          details:
-            document?.files?.map(({ mimeType, fileName, variant, fileId, imageUrl }) => {
-              const title = generateDocumentTitle({
-                category: document?.category ?? '',
-                type: document?.type ?? '',
-                variant,
-              });
-
-              return {
-                id: fileId,
-                title,
-                fileType: mimeType,
-                fileName,
-                imageUrl,
-              };
-            }) ?? [],
-        })) ?? [];
-
-      return [
-        ...adaptedDocumentsV2,
-        ...(identificationDocuments?.map((document, documentIndex) => ({
-          ...document,
-          details: documentPagesToDetailsAdapter({
-            document,
-            documentIndex,
-          }),
-        })) ?? []),
-      ];
-    }
-
-    return passedDocuments?.map((document, documentIndex) => ({
-      ...document,
-      details: documentPagesToDetailsAdapter({
-        document,
-        documentIndex,
-      }),
-    }));
-  };
-
-  const documents = getDocuments();
-
-  const issuerCountryCode = extractCountryCodeFromDocuments(documents ?? []);
-  const documentsSchemas = getDocumentsSchemas(issuerCountryCode, workflow);
-
-  return {
-    documents,
-    documentsSchemas,
-    isLoading: isLoadingDocumentsV2 || storageFilesQueryResult?.some(({ isLoading }) => isLoading),
-  };
-};
+import { FunctionComponent, useCallback } from 'react';
+import { toTitleCase } from 'string-ts';
 
 export const useDocumentBlocks = ({
   workflow,
@@ -371,7 +212,7 @@ export const useDocumentBlocks = ({
                     Re-upload needed
                     {!isLegacyReject && (
                       <X
-                        className="h-4 w-4 cursor-pointer"
+                        className="size-4 cursor-pointer"
                         onClick={() =>
                           onMutateRemoveDecisionById({
                             documentId: id,
