+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/DocumentField/hooks/useDocumentUpload/helpers/get-file-or-fileid-from-documents-list/get-file-or-fileid-from-documents-list.ts
@@ -16,7 +16,6 @@ export const getFileOrFileIdFromDocumentsList = (
   }
 
   const filePath = composePathToFileId(documentIndex, pageProperty, pageIndex);
-  console.log('filePath', filePath);
   const fileOrFileId = get(documentsList, filePath, undefined);
 
   return fileOrFileId;
