--- packages/ui/src/components/organisms/Form/DynamicForm/fields/EntityFieldGroup/utils/delay.ts
@@ -1,1 +1,0 @@
-export const delay = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
