+++ packages/ui/src/components/organisms/Form/DynamicForm/hooks/external/useRequired/helpers/check-if-required/check-if-required.ts
@@ -26,11 +26,7 @@ export const checkIfRequired = (
             ]).every(result => result.result)
           : true;
 
-        if (!shouldValidate) {
-          return false;
-        }
-
-        return true;
+        return shouldValidate;
       })
     : false;
