+++ services/workflows-service/src/app.module.ts
@@ -49,7 +49,6 @@ import { RuleEngineModule } from './rule-engine/rule-engine.module';
 import { NotionModule } from '@/notion/notion.module';
 import { SecretsManagerModule } from '@/secrets-manager/secrets-manager.module';
 import { NoteModule } from '@/note/note.module';
-import { DocumentModule } from './document/document.module';
 import { MerchantMonitoringModule } from './merchant-monitoring/merchant-monitoring.module';
 
 export const validate = async (config: Record<string, unknown>) => {
@@ -136,7 +135,6 @@ export const validate = async (config: Record<string, unknown>) => {
     RuleEngineModule,
     NotionModule,
     SecretsManagerModule,
-    DocumentModule,
   ],
   providers: [
     {
