+++ services/workflows-service/src/common/utils/get-entity-id/get-entity-id.ts
@@ -1,0 +1,13 @@
+import { BadRequestException } from '@nestjs/common';
+
+export const getEntityId = (data: { businessId?: string; endUserId?: string }) => {
+  if (data.businessId) {
+    return data.businessId;
+  }
+
+  if (data.endUserId) {
+    return data.endUserId;
+  }
+
+  throw new BadRequestException('Business or end user id is required');
+};
