+++ services/workflows-service/src/storage/storage.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { FileRepository } from './storage.repository';
 import { IFileIds } from './types';
 import { File, Prisma } from '@prisma/client';
@@ -18,7 +18,6 @@ import { z } from 'zod';
 import { HttpService } from '@nestjs/axios';
 import { AppLoggerService } from '@/common/app-logger/app-logger.service';
 import { readFileSync } from 'fs';
-import { randomUUID } from 'crypto';
 
 @Injectable()
 export class StorageService {
